SME02/requirements.txt
# Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
jinja2==3.1.4
python-multipart==0.0.9

# AI / LLM
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# PDF Generation
reportlab==4.2.2
xhtml2pdf==0.2.17

# Utilities
aiofiles==24.1.0
PyMuPDF>=1.24.0
python-docx==1.1.0
markdown==3.10.2
pypdf==5.1.0
duckduckgo-search==7.3.2

# Vector RAG Upgrade
chromadb<0.6.0; python_version < "3.13"
langchain-chroma==0.1.4; python_version < "3.13"
tiktoken==0.7.0; python_version < "3.13"
